+++ src/model/User.ts
@@ -4,5 +4,4 @@ export interface User{
     firstName: string,
     lastName: string,
     username: string,
-
 }
