+++ src/services/UserService.ts
@@ -1,7 +1,8 @@
 import {User} from "../model/User";
 import axios, {AxiosResponse} from 'axios';
+import {baseURL} from "./EnvironmentService";
 
 // User API calls to our server separated from component logic.
 export function getAllUsers(): Promise<AxiosResponse<User[]>> {
-    return axios.get<User[]>("http://localhost:8080/users");
+    return axios.get<User[]>(`${baseURL}/users`);
 }
